refactor(collage): Collage import 경로
